Changes to Game:StartStageTransition() & Level:SetStage() (#399)
* Correct StartStageTransition Bug

Corrects the bug listed on StartStageTransition, Originally stated that defining a player would cause the game to crash. The opposite is in fact the case.

* Add explanation of SetStage()

Added description of how the arguments passed to SetStage() work. May need rewording

* More clarity on StageTypeOffset

Added another line in the description of StageTypeOffset to show that floors 1 and 2 have the same starting value

* Added StartStageTransition inconsistency warning

Added a small section to the Bug of StartStageTransition to warn that even when used correctly, it is inconsistent and may crash for unknown reasons

* Updated formatting of Game.md changes

* Updated formatting of Level.md changes

* Reworded StartStageTransition desc for clarity

* Removed Transition Types, seemed to have no effect

* Added link to stages.xml

* Replaced incorrect desc from official docs + more

* Rearranged things for readability

* Added note on apparent lack of use for Animation